Refactor TIMDEX models
Why are these changes being introduced:

* The timdex model was handling interactions for both search and
  retrieve, which are separate interactions that don't intersect other
  than the client setup

Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/RDI-118

How does this address that need:

* The shared timdex graphql client logic is moved to a timdex_base class
  and timdex_search and timdex_record are created.

Side Effects:

There may be some confusion calling the new model `timdex_record` rather than `timdex_retrieve`. However, other parts of this app are already referring to a single record retrieval as `record` and not `retrieve` so for now I think this is the correct name.
