+++ app/models/timdex_record.rb
@@ -1,20 +1,8 @@
 require 'graphql/client'
 require 'graphql/client/http'
 
-class Timdex
-  HTTP = GraphQL::Client::HTTP.new(ENV.fetch('TIMDEX_GRAPHQL', '')) do
-    def headers(*)
-      {
-        'User-Agent': 'MIT Libraries Client'
-      }
-    end
-  end
-
-  Schema = GraphQL::Client.load_schema('config/schema/schema.json')
-
-  Client = GraphQL::Client.new(schema: Schema, execute: HTTP)
-
-  RecordQuery = Timdex::Client.parse <<-'GRAPHQL'
+class TimdexRecord < TimdexBase
+  Query = TimdexBase::Client.parse <<-'GRAPHQL'
     query($id: String!) {
       recordId(id: $id) {
         alternateTitles {
@@ -103,62 +91,6 @@ class Timdex
         summary
         timdexRecordId
         title
-      }
-    }
-  GRAPHQL
-
-  SearchQuery = Timdex::Client.parse <<-'GRAPHQL'
-    query($q: String!) {
-      search(searchterm: $q) {
-        hits
-        records {
-          timdexRecordId
-          title
-          contentType
-          contributors {
-            kind
-            value
-          }
-          publicationInformation
-          dates {
-            kind
-            value
-          }
-          notes {
-            kind
-            value
-          }
-        }
-        aggregations {
-          contentFormat {
-            key
-            docCount
-          }
-          contentType {
-            key
-            docCount
-          }
-          contributors {
-            key
-            docCount
-          }
-          languages {
-            key
-            docCount
-          }
-          literaryForm {
-            key
-            docCount
-          }
-          source {
-            key
-            docCount
-          }
-          subjects {
-            key
-            docCount
-          }
-        }
       }
     }
   GRAPHQL
