+++ test/vcr_cassettes/timdex_1234-5678.yml
@@ -5,14 +5,14 @@ http_interactions:
     uri: https://FAKE_TIMDEX_HOST/graphql
     body:
       encoding: UTF-8
-      string: '{"query":"query Timdex__SearchQuery($q: String!) {\n  search(searchterm:
+      string: '{"query":"query TimdexSearch__Query($q: String!) {\n  search(searchterm:
         $q) {\n    hits\n    records {\n      timdexRecordId\n      title\n      contentType\n      contributors
         {\n        kind\n        value\n      }\n      publicationInformation\n      dates
         {\n        kind\n        value\n      }\n      notes {\n        kind\n        value\n      }\n    }\n    aggregations
         {\n      contentFormat {\n        key\n        docCount\n      }\n      contentType
         {\n        key\n        docCount\n      }\n      contributors {\n        key\n        docCount\n      }\n      languages
         {\n        key\n        docCount\n      }\n      literaryForm {\n        key\n        docCount\n      }\n      source
-        {\n        key\n        docCount\n      }\n      subjects {\n        key\n        docCount\n      }\n    }\n  }\n}","variables":{"q":"1234-5678"},"operationName":"Timdex__SearchQuery"}'
+        {\n        key\n        docCount\n      }\n      subjects {\n        key\n        docCount\n      }\n    }\n  }\n}","variables":{"q":"1234-5678"},"operationName":"TimdexSearch__Query"}'
     headers:
       Accept-Encoding:
       - gzip;q=1.0,deflate;q=0.6,identity;q=0.3
@@ -30,7 +30,7 @@ http_interactions:
       Server:
       - Cowboy
       Date:
-      - Thu, 26 May 2022 21:02:06 GMT
+      - Thu, 26 May 2022 17:13:17 GMT
       Connection:
       - keep-alive
       X-Frame-Options:
@@ -54,9 +54,9 @@ http_interactions:
       Cache-Control:
       - max-age=0, private, must-revalidate
       X-Request-Id:
-      - 8c7d69e4-66bb-4f04-b505-e390709bf6f0
+      - 8df91947-07ea-4d7d-8007-41253df3de4c
       X-Runtime:
-      - '0.070215'
+      - '0.080935'
       Strict-Transport-Security:
       - max-age=63072000; includeSubDomains
       Transfer-Encoding:
@@ -66,5 +66,5 @@ http_interactions:
     body:
       encoding: UTF-8
       string: '{"data":{"search":{"hits":0,"records":[],"aggregations":{"contentFormat":[],"contentType":[],"contributors":[],"languages":[],"literaryForm":[],"source":[],"subjects":[]}}}}'
-  recorded_at: Thu, 26 May 2022 21:02:06 GMT
+  recorded_at: Thu, 26 May 2022 17:13:17 GMT
 recorded_with: VCR 6.1.0
